reformatted code and moved unused functions to a new class for future work
