+++ src/main/kotlin/presentation/FoodChangeMoodConsole.kt
@@ -17,46 +17,19 @@ class FoodChangeMoodConsole() {
         showOptions()
         val userChoice = getUserChoice()
         when (userChoice) {
-            0 -> {
-                sayGoodBye()
-                return
-            }
-
-            1 -> {
-                explainFirstChoice()
-
-            }
-
-            3 -> {
-                explainThirdChoice()
-
-            }
-
-            4 -> {
-                explainFourthChoice()
-
-            }
-
-            5 -> {
-                explainFifthChoice()
-                guessGame()
-            }
-
-            6 -> {
-                explainSixthChoice()
-                noEggsSweet()
-            }
-
-            7 -> {
-                explainSeventhChoice()
-
-            }
-
-            else -> {
-                println(
-                    "${ConsoleColors.RED_COLOR}Invalid Choice!${ConsoleColors.RESETCOLOR}\n" +
-                            "We'll support other features in the future!"
-                )
+            0 -> { sayGoodBye()
+                    return
+            }1 -> { explainFirstChoice()
+            }3 -> { explainThirdChoice()
+            }4 -> { explainFourthChoice()
+            }5 -> { explainFifthChoice()
+                    guessGame()
+            }6 -> { explainSixthChoice()
+                    noEggsSweet()
+            }7 -> { explainSeventhChoice()
+            }else -> {
+                println( "${ConsoleColors.RED_COLOR}Invalid Choice!${ConsoleColors.RESETCOLOR}\n" +
+                            "We'll support other features in the future!")
             }
         }
         chooseOption()
@@ -169,14 +142,6 @@ class FoodChangeMoodConsole() {
         )
     }
 
-    private fun explainSecondChoice() {
-        println(
-            "enter a country name, either in capital or small or even a mix. " +
-                    "don't worry about spelling mistakes! " +
-                    "and we will show all meals in our list related to that country"
-        )
-    }
-
     private fun explainThirdChoice() {
         println("a list of Iraqi meals will be shown. you'll see any meal related to Iraq in our list")
     }
@@ -210,61 +175,6 @@ class FoodChangeMoodConsole() {
                     "you can either like it (to view full details) or dislike it (to get another Keto meal)."
         )
     }
-
-    private fun explainEighthChoice() {
-        println("enter a date, and we'll show a list of meals that were added in that date")
-    }
-
-    private fun explainNinthChoice() {
-        println(
-            "enter a desired amount of calories and protein, " +
-                    "and we will give you a list of meals that match or approximate those values."
-        )
-    }
-
-    private fun explainTenthChoice() {
-        println(
-            "enter a country name, and we'll return " +
-                    "up to 20 randomly ordered meals related to that country."
-        )
-    }
-
-    private fun explainEleventhChoice() {
-        println(
-            "we'll Display a meal name and three ingredient options (one correct, two incorrect)." +
-                    " you'll guess once. A correct guess earns 1000 points; an incorrect guess ends the game. " +
-                    "The game also ends after 15 correct answers. we'll Display the final score at the end."
-        )
-    }
-
-    private fun explainTwelfthChoice() {
-        println(
-            "for potato lovers! we'll Show a random list of 10 meals that include potatoes " +
-                    "in their ingredients."
-        )
-    }
-
-    private fun explainThirteenthChoice() {
-        println(
-            "we'll Suggest a meal with more than 700 calories. you can either like it (to view full details)" +
-                    " or dislike it (to get another meal with high calories). "
-        )
-    }
-
-    private fun explainFourteenthChoice() {
-        println(
-            "we'll Show a list of all seafood meals sorted by protein content, from highest to lowest." +
-                    " we'll display the rank (starting from 1), meal name, and protein amount."
-        )
-    }
-
-    private fun explainFifteenthChoice() {
-        println(
-            "if u and A large group of friends are traveling to Italy," +
-                    " we'll show you a list of original Italian dishes, that are suitable for large groups."
-        )
-    }
-
     private fun showOptions() {
         println(
             "you can choose one of the following:" +
